requirements.txt
streamlit
python-dotenv
transformers
torch
huggingface_hub
sentencepiece
